# swagger.yaml
openapi: 3.0.0
info:
  description: |
    Pinterest CM sv-cl API
  version: 1.2.1
  title: Gears Of Peace
  license:
    name: Gears Of Peace
    url: https://github.com/tut_budet_ssilka_na_front
servers:
  - url: https:/zaregayu_pozdnee
    description: Main production server
...
paths:
...
  /auth/login:
    post:
      tags:
        - auth
      summary: Logs user into the system
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
        required: true
      responses:
        '204':
          description: Successful operation
        '400':
          description: Invalid username or password supplied
        '401':
          description: Wrong password
        '403':
          description: You are already authorized. Log out first
        '404':
          description: User not found
  /auth/logout:
    post:
      operationId: logoutUser
      tags:
        - auth
      summary: Log out from current session
      responses:
        '204':
          description: Logged out
        '401':
          description: User unauthorized
...
  /shop:
    post:
      operationId: createShop
      tags:
        - shop
      summary: Create new shop
      description: This can only be done by authorized user
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                shopInfo:
                  type: object
                  properties:
                    managerIDs:
                      type: array
                      items:
                        type: integer
                        format: int
                    name:
                      type: string
                    description:
                      type: string
                shopImage:
                  type: string
                  format: binary
        description: New product
        required: true
      responses:
        '201':
          description: Successfully created shop
          content:
            application/json:
              schema:
                properties:
                  ID:
                    type: integer
        '400':
          description: Failed to create shop due to invalid data
        '403':
          description: Can't create shop while not logged in as manager
  /shop/{shopID}:
    get:
      operationId: getShopByID
      tags:
        - shop
      summary: Get shop by ID
      parameters:
        - name: shopID
          in: path
          schema:
            type: integer
            format: int
          description: The ID that needs to be fetched
          required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shop'
        '400':
          description: Invalid shop ID supplied
        '404':
          description: Shop not found
    delete:
      operationId: deleteShopByID
      tags:
        - shop
      summary: Delete shop by ID
      description: This can only be done by manager
      parameters:
        - name: shopID
          in: path
          schema:
            type: integer
            format: int
          description: The ID of shop that needs to be deleted
          required: true
      responses:
        '204':
          description: Successful operation
        '400':
          description: Invalid shop ID supplied
        '401':
          description: User unauthorized as manager
        '404':
          description: Shop not found
components:
  schemas:
...
    Shop:
      type: object
      properties:
        ID:
          type: integer
          format: int
        managerIDs:
          type: array
          items:
            type: integer
            format: int
        name:
          type: string
        description:
          type: string
        avatarLink:
          type: string
      required:
        - ID
        - title
...